translation/requirements.txt
fastapi
uvicorn
transformers
torch
textwrap
pymupdf
python-docx
openpyxl
python-pptx
